+++ apps/remnant2toolkit/app/tracker/item-list.tsx
@@ -22,6 +22,7 @@ import {
   getFilteredItemsForCategory,
   parseUrlFilters,
 } from '@/app/(components)/filters/item-tracker/utils'
+import { modItems } from '@/app/(data)/items/mod-items'
 import { Item } from '@/app/(data)/items/types'
 import {
   ItemTrackerLocalStorage,